# Conflicts:
#	ui/package-lock.json
#	ui/src/app/App.tsx
#	ui/src/app/screens/productsFlow/pricing/Pricing.tsx
#	ui/src/app/screens/styledComponents/navBars.tsx

+++ ui/copyProtos.sh
@@ -1,4 +1,6 @@
-rm -r protos
+if [ -d "protos" ]; then
+  rm -r protos
+fi
 if [ -d "src/generated" ]; then
   rm -r src/generated
 fi
@@ -8,5 +10,6 @@ rsync -rR ../gateway/src/main/proto/./com/example/gateway/**/*.proto protos
 rsync -rR ../common/src/main/proto/./com/example/common/**/*.proto protos
 rsync -rR ../common/src/main/proto/./com/example/common/*.proto protos
 rsync -rR ../gateway/target/protobuf_external/./**/*.proto protos
-rsync -rR ../gateway/target/protobuf_external/./**/**/*.proto protos
+rsync -rR ../gateway/target/protobuf_external_src/./**/*.proto protos
+rsync -rR ../gateway/target/protobuf_external_src/./**/**/*.proto protos
 rm protos/google/api/service.proto
